Plant-Doctor/requirements.txt

streamlit==1.36.0
tensorflow==2.16.1
Pillow==10.4.0
numpy==1.26.4
pandas==2.2.2
